Refactor file transfer and disk free space utilities
Move run_file_transfer and run_virtio_serial_file_transfer from
utils_test/__init__.py into new utils_test/file_transfer.py module.
Re-exports in __init__.py preserve backward compatibility.

Convert utils_disk.py to utils_disk/ package. Move get_free_disk
from utils_misc.py to utils_disk/free_space.py and add new
check_free_disk helper. Re-exports in utils_disk/__init__.py and
utils_misc.py preserve backward compatibility.

Add guest disk free space validation in run_file_transfer to fail
early with a clear error when the VM has insufficient space for
the transfer.
